+++ src/app/routes/my-network.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-useHead({
+useSeoMeta({
   title: "My Network",
 });
 </script>
