+++ src/entities/game-level/api/index.ts
@@ -1,1 +1,2 @@
+export type { GameLevelShortInfoDto } from "./interfaces/short-info.dto";
 export { GameLevelService } from "./service";
