+++ src/entities/game-level/api/interfaces/game-levels.req.dto.ts
@@ -1,5 +1,3 @@
-import type { GameLevelShortInfoDto } from "./short-info.dto";
-
 export interface GameLevelsReqDto {
   info: GameLevelShortInfoDto[];
 }
