+++ src/entities/game-level/ui/card/enums/states.ts
@@ -1,0 +1,8 @@
+export enum GameLevelCardStates {
+  Error = "error",
+  Awaiting = "awaiting",
+  Active = "active",
+  Freeze = "freeze",
+  Unavailable = "unavailable",
+  Default = "default",
+}
