+++ src/entities/game-level/ui/card/interfaces/index.ts
@@ -1,0 +1,1 @@
+export type { GameLevelCardEntryProps } from "./props";
