+++ src/entities/game-level/ui/card/interfaces/props.ts
@@ -1,0 +1,3 @@
+export interface GameLevelCardEntryProps {
+  info: GameLevelShortInfoDto;
+}
