+++ src/entities/game-level/ui/card/states/active/interfaces/index.ts
@@ -1,0 +1,1 @@
+export type { GameLevelCardActiveProps } from "./props";
