+++ src/entities/game-level/ui/card/states/active/interfaces/props.ts
@@ -1,0 +1,3 @@
+export interface GameLevelCardActiveProps {
+  info: GameLevelShortInfoDto;
+}
