+++ src/entities/game-level/ui/card/states/awaiting/interfaces/index.ts
@@ -1,0 +1,1 @@
+export type { GameLevelCardAwaitingProps } from "./props";
