+++ src/entities/game-level/ui/card/states/awaiting/interfaces/props.ts
@@ -1,0 +1,3 @@
+export interface GameLevelCardAwaitingProps {
+  info: GameLevelShortInfoDto;
+}
