+++ src/entities/game-level/ui/card/states/default/interfaces/index.ts
@@ -1,0 +1,1 @@
+export type { GameLevelCardDefaultProps } from "./props";
