+++ src/entities/game-level/ui/card/states/default/interfaces/props.ts
@@ -1,0 +1,3 @@
+export interface GameLevelCardDefaultProps {
+  info: GameLevelShortInfoDto;
+}
