+++ src/entities/game-level/ui/card/states/error/interfaces/index.ts
@@ -1,0 +1,1 @@
+export type { GameLevelCardErrorProps } from "./props";
