+++ src/entities/game-level/ui/card/states/error/interfaces/props.ts
@@ -1,0 +1,3 @@
+export interface GameLevelCardErrorProps {
+  info: GameLevelShortInfoDto;
+}
