+++ src/entities/game-level/ui/card/states/freeze/interfaces/index.ts
@@ -1,0 +1,1 @@
+export type { GameLevelCardFreezeProps } from "./props";
