+++ src/entities/game-level/ui/card/states/freeze/interfaces/props.ts
@@ -1,0 +1,3 @@
+export interface GameLevelCardFreezeProps {
+  info: GameLevelShortInfoDto;
+}
