+++ src/entities/game-level/ui/card/states/unavailable/interfaces/index.ts
@@ -1,0 +1,1 @@
+export type { GameLevelCardUnavailableProps } from "./props";
