+++ src/entities/game-level/ui/card/states/unavailable/interfaces/props.ts
@@ -1,0 +1,3 @@
+export interface GameLevelCardUnavailableProps {
+  info: GameLevelShortInfoDto;
+}
