--- src/entities/game-level/ui/enums/index.ts
@@ -1,1 +1,0 @@
-export { GameLevelStates } from "./states";
