--- src/entities/game-level/ui/enums/states.ts
@@ -1,7 +1,0 @@
-export enum GameLevelStates {
-  Error = "error",
-  Awaiting = "awaiting",
-  Active = "active",
-  Freeze = "freeze",
-  Unavailable = "unavailable",
-}
