--- src/entities/game-level/ui/interfaces/index.ts
@@ -1,1 +1,0 @@
-export type { GameLevelEntryProps } from "./props";
