--- src/entities/game-level/ui/interfaces/props.ts
@@ -1,5 +1,0 @@
-import type { GameLevelShortInfoDto } from "../../api/interfaces";
-
-export interface GameLevelEntryProps {
-  info: GameLevelShortInfoDto;
-}
