--- src/entities/game-level/ui/states/active/interfaces/index.ts
@@ -1,1 +1,0 @@
-export type { GameLevelActiveProps } from "./props";
