--- src/entities/game-level/ui/states/awaiting/interfaces/index.ts
@@ -1,1 +1,0 @@
-export type { GameLevelAwaitingProps } from "./props";
