--- src/entities/game-level/ui/states/awaiting/interfaces/props.ts
@@ -1,5 +1,0 @@
-import type { GameLevelShortInfoDto } from "~/src/entities/game-level/api/interfaces";
-
-export interface GameLevelAwaitingProps {
-  info: GameLevelShortInfoDto;
-}
