--- src/entities/game-level/ui/states/default/interfaces/index.ts
@@ -1,1 +1,0 @@
-export type { GameLevelDefaultProps } from "./props";
