--- src/entities/game-level/ui/states/default/interfaces/props.ts
@@ -1,5 +1,0 @@
-import type { GameLevelShortInfoDto } from "~/src/entities/game-level/api/interfaces";
-
-export interface GameLevelDefaultProps {
-  info: GameLevelShortInfoDto;
-}
