--- src/entities/game-level/ui/states/error/interfaces/index.ts
@@ -1,1 +1,0 @@
-export type { GameLevelErrorProps } from "./props";
