--- src/entities/game-level/ui/states/freeze/interfaces/index.ts
@@ -1,1 +1,0 @@
-export type { GameLevelFreezeProps } from "./props";
