--- src/entities/game-level/ui/states/freeze/interfaces/props.ts
@@ -1,5 +1,0 @@
-import type { GameLevelShortInfoDto } from "~/src/entities/game-level/api/interfaces";
-
-export interface GameLevelFreezeProps {
-  info: GameLevelShortInfoDto;
-}
