--- src/entities/game-level/ui/states/unavailable/interfaces/index.ts
@@ -1,1 +1,0 @@
-export type { GameLevelUnavailableProps } from "./props";
