--- src/entities/game-level/ui/states/unavailable/interfaces/props.ts
@@ -1,5 +1,0 @@
-import type { GameLevelShortInfoDto } from "~/src/entities/game-level/api/interfaces";
-
-export interface GameLevelUnavailableProps {
-  info: GameLevelShortInfoDto;
-}
