+++ src/features/game-level/ui/modal/interfaces/props.ts
@@ -1,3 +1,3 @@
 export interface GameLevelModalProps {
-  idCard: number;
+  idLevel: number;
 }
