+++ src/pages/my-network/index.vue
@@ -1,7 +1,3 @@
-<script setup lang="ts">
-import { UserInfoWidget } from "#components";
-</script>
-
 <template>
   <div class="mt-37.5 max-md:mt-8">
     <UserInfoWidget class="mb-14.25 max-md:mb-6" />
